#12401. Add needed enrolment status check to listener.
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@17327 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ common/common-services/src/main/java/ish/oncourse/services/lifecycle/EnrolmentListener.java
@@ -8,6 +8,7 @@ import org.apache.cayenne.annotation.PostUpdate;
 
 import ish.oncourse.model.Attendance;
 import ish.oncourse.model.Enrolment;
+import ish.oncourse.model.EnrolmentStatus;
 import ish.oncourse.model.Outcome;
 import ish.oncourse.model.Session;
 import ish.oncourse.services.persistence.ICayenneService;
@@ -16,41 +17,45 @@ import ish.oncourse.services.persistence.ICayenneService;
  * Removes Attendances and Outcomes for Enrolment if its status becomes unsuccessful.
  * 
  * @author dzmitry
- *
  */
 public class EnrolmentListener {
-	
+
 	/**
 	 * Cayenne service.
 	 */
 	private ICayenneService cayenneService;
-	
+
 	public EnrolmentListener(ICayenneService cayenneService) {
 		super();
 		this.cayenneService = cayenneService;
 	}
-	
+
 	@PostUpdate(value = Enrolment.class)
 	public void postUpdate(Enrolment e) {
-		ObjectContext context = cayenneService.newContext();
 		
-		Enrolment enrol = (Enrolment) context.localObject(e.getObjectId(), null);
-		
-		if (enrol != null) {
-			for (Session session : enrol.getCourseClass().getSessions()) {
-				Attendance a = enrol.getAttendanceForSessionAndStudent(session, enrol.getStudent());
-				if (a != null) {
-					context.deleteObject(a);
+		if (EnrolmentStatus.CANCELLED.equals(e.getStatus()) || EnrolmentStatus.FAILED.equals(e.getStatus())
+				|| EnrolmentStatus.REFUNDED.equals(e.getStatus())) {
+
+			ObjectContext context = cayenneService.newContext();
+			Enrolment enrol = (Enrolment) context.localObject(e.getObjectId(), null);
+
+			if (enrol != null) {
+				// removing attendances
+				for (Session session : enrol.getCourseClass().getSessions()) {
+					Attendance a = enrol.getAttendanceForSessionAndStudent(session, enrol.getStudent());
+					if (a != null) {
+						context.deleteObject(a);
+					}
 				}
+
+				// removing outcomes
+				List<Outcome> outcomes = new ArrayList<Outcome>(enrol.getOutcomes());
+				for (Outcome o : outcomes) {
+					context.deleteObject(o);
+				}
+
+				context.commitChanges();
 			}
-			
-			List<Outcome> outcomes = new ArrayList<Outcome>(enrol.getOutcomes());
-			for (Outcome o : outcomes) {
-				context.deleteObject(o);
-			}
-			
-			context.commitChanges();
 		}
 	}
-
 }
